created an items container
