+++ src/pages/Home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Items from '../components/items/Items';
+import HomePageItems from '../components/items/HomePageItems';
 import Footer from '../components/common/Footer';
 import CustomTitle from "../components/common/CustomTitle";
 
@@ -7,7 +7,7 @@ export default function Home() {
     return (
         <div>
             <CustomTitle>Featured Deals</CustomTitle>
-            <Items />
+            <HomePageItems />
             <Footer />
         </div>
     )
